Fix for large files (It is the typical bug of many kioslaves; the createUDSEntry function uses a "long int" as size instaed of the correct "long long int".) (Foreward port of revision 468882; might not compile)
svn path=/trunk/KDE/kdesdk/kioslave/; revision=468883

+++ svn/svn.cpp
@@ -477,7 +477,7 @@ void kio_svnProtocol::listDir(const KURL& url){
 	svn_pool_destroy (subpool);
 }
 
-bool kio_svnProtocol::createUDSEntry( const QString& filename, const QString& user, long int size, bool isdir, time_t mtime, UDSEntry& entry) {
+bool kio_svnProtocol::createUDSEntry( const QString& filename, const QString& user, long long int size, bool isdir, time_t mtime, UDSEntry& entry) {
 	kdDebug(7128) << "MTime : " << ( long )mtime << endl;
 	kdDebug(7128) << "UDS filename : " << filename << endl;
 	UDSAtom atom;
